requirements.txt
# Hera runtime dependencies
# Install with: pip install -r requirements.txt

# Qt GUI framework (widgets, signals, painting)
PySide6==6.11.0

# HTTP client — REST API calls to Hercules
requests==2.32.5

# Image loading — device bitmaps in the room view
Pillow==12.1.1

# PDF generation — printer (1403) workspace PDF export
fpdf2==2.8.7
